Create standart functions

+++ src/mySimpleComputer.h
@@ -6,7 +6,7 @@
 #define ARCHITECTURE_MYSIMPLECOMPUTER_H
 extern const int mem_size;
 extern int flag;
-extern int memory[];
+extern short memory[];
 int sc_memoryInit();
 int sc_memorySet(int address, int value);
 int sc_memoryGet(int address, int* value);
